backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
netaddr==0.9.0
dnspython==2.4.2
pyyaml==6.0.1
jinja2==3.1.2
django-auditlog==2.3.0
djangorestframework-api-key==3.0.0
django-simple-history==3.4.0
django-import-export==3.3.5